+++ src/lib/components/ui/empty-state/index.ts
@@ -1,0 +1,7 @@
+import Root from './empty-state.svelte';
+
+export {
+	Root as EmptyState,
+	//
+	Root
+};