+++ src/routes/(app)/[budgetId=id]/accounts/new/+page.svelte
@@ -2,6 +2,12 @@
 	import { AccountCreate } from '$lib/components/features/account';
 	import * as Page from '$lib/components/ui/page';
 	import { m } from '$lib/paraglide/messages';
+	import { getBudget } from '$lib/remote-functions/budget.remote';
+	import { getBudgetId } from '$lib/utils/budget-id-context';
+
+	const budgetId = getBudgetId();
+
+	const { currency } = $derived(await getBudget(budgetId()));
 </script>
 
 <Page.Root class="max-w-lg">
@@ -16,6 +22,6 @@
 	</Page.Header>
 
 	<Page.Content>
-		<AccountCreate />
+		<AccountCreate {currency} />
 	</Page.Content>
 </Page.Root>